+++ src/Http/Controllers/FrontendController.php
@@ -18,11 +18,9 @@ class FrontendController extends FrontendBaseController
      *
      * Render page (@override)
      * @param Request $request
-     * @param array $infoKeeper
      * @return array|\Illuminate\Contracts\View\Factory|\Illuminate\View\View
      */
     public function index(Request $request) {
-        //return $request->infoKeeper;
         return parent::index($request);
 
     }