+++ src/User.php
@@ -18,9 +18,7 @@ class User extends Authenticatable
      *
      * @var array
      */
-    protected $fillable = [
-        'name', 'email', 'password','user_type'
-    ];
+    protected $guarded = array();
 
     /**
      * The attributes that should be hidden for arrays.